web/pyproject.toml
[tool.poetry]
name = "rattletat.com"
version = "0.1.0"
description = ""
authors = ["Michael Brauweiler <michael.brauweiler@posteo.de>"]

[tool.poetry.dependencies]
python = "^3.8"
django-environ = "^0.4.5"
django-model-utils = "^4.0.0"
argon2_cffi = "^20.1.0"
bleach = "^3.1.5"
django-next-prev = "^1.0.1"
fabric2 = "^2.5.0"
pillow = "^7.2.0"
django-markdownify = "^0.8.0"
markdown-checklist = "^0.4.3"
python-memcached = "^1.59"
django-htmlmin = "^0.11.0"
psycopg2-binary = "^2.8.6"
django-admin-sortable2 = "^0.7.8"

[tool.poetry.dev-dependencies]
ipython = "^7.15.0"
flake8 = "^3.8.2"
black = "^19.10b0"
django-debug-toolbar = "^2.2"
isort = "^4.3.21"
fabric = "^2.5.0"
django-compressor = "^2.4"
pylint = "^2.7.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
